refactor(leaderboard): separated CarDataBuilder class  into a new module. test: update import line in car_data_builder tests module
